+++ src/main/java/com/hustleind/service/RequestService.java
@@ -6,7 +6,7 @@ import org.springframework.util.MultiValueMap;
 import java.util.List;
 
 public interface RequestService {
-    Request getRequestById(int id);
+    Request getRequestById(Long id);
     List<Request> getAllRequests();
     boolean addRequest(Request request);
     boolean editRequest(Request request);
